[move-ide] Added an option to force bundled move-analyzer
## Description 

I recently encountered a user problem when their VSCode was picking up
their on-path `haneul` binary to get `move-analyzer` functionality out of
it but it did not work. What did work was to keep using the bundled
`move-analyzer` but there wasn't a convenient option to force it. This
PR adds this kind of option to the extension's config.

Generally, we want people to use `move-analyzer` from `haneul` binary if
available so that they have consistent compiler behavior (mostly with
respect to diagnostics), That's why the new option defaults to `false`.

Another use for this option is if someone is savvy enough to have a
custom `haneul` binary on their path but it's a debug version - we
definitely don't want `move-analyzer` running in debug mode

## Test plan 

Tested manually that flipping the flag back and forth results in
expected behavior