rpc-alt: ObjectFilter.MatchNone
## Description

Implement `MatchNone` for owned object filters. This is implemented by
running a plain owned object query and then applying the more complex
filter in memory as a sequential scan.

We do it this way to avoid giving the postgres query planner a chance to
disappoint us -- it is also similar to how the same query is implemented
on fullnodes (where the DB is embedded).

## Test plan

New E2E tests:

```
haneul$ cargo nextest run -p haneul-indexer-alt-e2e-tests
```