[bridge] add lbtc vault monitoring
## Description 

A PR to enable LBTC monitoring. williampsmith we'll want to add this to
one of our nodes for the LBTC launch. Because LBTC (token ID 6) is not
added to testnet, we'll need to either enable LBTC to testnet (no code
change required) before landing this PR OR come up with an alternative
(maybe only push metrics for a token ID if its actually enabled? right
now address of token ID 6 on testnet will return 0x0 so may break node.
anyways not going to land this until we decide (after offsite)

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------