[rpc-alt] fix fullnode rpc url arg
## Description 

I realized the way we handled the write arg (fullnode rpc url) was
incorrect. We wanted it to be an optional arg but it's not actually
optional due to the `#[command(flatten)]` attribute and causes this
error when a fullnode url is not provided:
```
error: the following required arguments were not provided:
  <FULLNODE_RPC_URL>

Usage: haneul-indexer-alt-jsonrpc rpc --database-url <DATABASE_URL> --config <CONFIG> <FULLNODE_RPC_URL>
```

## Test plan 

edited tests

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: