[move] Add `move-vm-types` to execution version cuts and cut historical versions
## Description 

This cuts the different versions of the move-vm-types, and updates the
tomls and execution layer script to work with them. You can ignore all
`.rs` files. The only interesting changes are in `.py` and `.toml`
files.

This also adds one small change to make the execution layer cut script
work with git workspaces (it previously assumed that the git root and
root of the repo were one and the same).

## Test plan 

CI