[move-model-2] Reorganize and add shared interface
## Description 

- I did not fully anticipate needing to write code that could work over
either the compiled model OR the source model--with specializations
depending on which info you have in hand
- My solution to this was a "marker" generic `HAS_SOURCE` which has
extra functions implemented when you have the source (and in one case
when you know you are compiled)
- The const generic ensures statically that you know which case you are
in
- There is a helper `kind` function that lets you case on which one you
are in

## Test plan 

- 👀  and clippy
- This is used in the haneul-client-gen rewrite
https://github.com/tnowacki/haneul-client-gen/tree/model2

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: