[clli] Remove console command in the CLI
## Description 

Remove the console command in the CLI as it's unlikely that it's being
used and it has not been maintained at all.

## Test plan 

Current tests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: Starting with v1.45.0, the `haneul console` command is not
available anymore.
- [ ] Rust SDK: