graphql-alt: initial scaffolding
## Description

Initial scaffolding for a new GraphQL service, based on
`haneul-indexer-alt-schema`. As of this change, the service contains only
the logic to start up and shutdown itself and an IDE, and a single dummy
query to test both with.

## Test plan

Run:
```
haneul$ cargo run -p haneul-indexer-alt-graphql -- rpc
```

And visit `http://localhost:7000/graphql`.