[Consensus] rework commit & block timestamps
## Description 

This PR is doing the following:
1. Is dropping the timestamp monotonicity block ancestor checks. We'll
no longer check whether a block's ancestors timestamps are strictly
lower than the block in question.
2. Calculating the commit's timestamp by using the leader's strong
ancestors timestamps median. The calculation is stake based.

As it has been discussed, this will make the protocol more robust when
it comes to block validity verification and repercussions of it. It is
also simplifying the block acceptance path.

## Test plan 

CI/PT

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------