[move/adapter] Make native extensions API fallible
## Description 

It always annoyed me that this API was infallible as it should really
return invariant violations if the extension is not found.

The bottom commit is the core of the change and the other two commits on
top are just the updates on the adapter side based on those changes.

## Test plan 

Make sure existing tests pass.