[move] Add static explicit marker for Move native extensions
## Description 

Replaces the usage of `TidAble<'a>` for `NativeExtensionMarker<'a>` in
the interface for native extensions. This makes trying to query for a
type in the native context extensions without registering the _exact_
type a static error. This explicitly does not have blanket
implementations for wrappers of an underlying type that implements the
trait unlike
[`TidAble`](https://docs.rs/better_any/latest/better_any/trait.TidAble.html).
This means you must register the exact type that you will be adding to
the extensions with the `NativeExtensionMarker` trait.

## Test plan 

Make sure existing tests pass.