store: introduce events_2 table to store events keyed by transaction digest
In Haneul TransactionEvents are not unique and multiple transactions can
produce the exact same set of TransactionEvents. Today these events are
stored in a table keyed by TransactionEventsDigest which can lead to an
unfortunate scenario when pruning is enabled. If we have two
transactions `A` and `B`, both of which produced events `E`, when the
pruner goes to prune transaction `A` it will also remove events `E` from
the database. This will lead to a situation where the node will be
unable to find locally the events produced by transaction `B`.

In order to fix this bug this patch introduces a new `events_2` table to
store events keyed by the transaction digest of the transaction that
produced them. It also updates the interface used to read events from
the db to require a `TransactionDigest` instead of a
`TransactionEventsDigest`. Internally the `events_2` table is first checked
but then there is fallback logic to lookup the events from the `events`
table. This fallback logic can eventually be removed once we've
completed a db migration and deprecate the `events` table.