graphql-alt: x-haneul-rpc-show-usage
## Description

This request header tells the GraphQL service to extend responses with
statistics on how much of each query limit was used.

I was also able to figure out what our previous issue with custom
headers was (it was that we were not making any effort to read a header
value), and with that fixed, we are once again able to use the
`Option<TypedHeader<T>>` extractor (rather than extracting all headers).

## Test plan

Manual testing -- making sure usage information is added to the response
when the header is present and is omitted when the header is omitted.