fix(rpc-alt): fix CoinMetadata and Display data-loader ordering
## Description

Both these data-loaders follow a pattern where they generate some probes
into the DB, based on the data loader key, form a set of these to query
the DB with and then later try to zip that set with the original
sequence of keys they were generated from.

If that sequence of keys is in a different order or contains duplicates,
then keys are going to end up mis-aligned with their equivalent probes.

## Test plan

:eyes:

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: