[indexer-alt]: Changes visibility of IngestionService and Indexer metrics
## Description 

Changes visibility of IngestionService and Indexer metrics to public.
This is useful for cases where the IngestionService is needed in a
standalone mode and out of the context of the Indexer.
For example, when a custom range of checkpoints need to be parsed.

```
let mut ingestion_service = IngestionService::new(
        client_args,
        IngestionConfig::default(),
        metrics,
        cancel.clone(),
    )?;
    let (mut rx, _) = ingestion_service.subscribe();

    // Fetch malicious transactions and start ingestion
    let checkpoint_ids = pipeline.fetch_malicious_trxs().await?;
    println!("Found {} checkpoint IDs to process", checkpoint_ids.len());

    let (_, _) = ingestion_service
        .run(checkpoint_ids)
        .await
        .context("Failed to start ingestion service")?;
```

## Test plan 

Existing tests pass

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: