Docs: Add note to functions only available on Devnet
## Description 

Add a note to the function docs of Haneul framework functions that are only
enabled on Devnet. Currently, this is oblivious to anyone reading the
docs.

## Test plan 

No functional changes, so no tests are needed.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------