Implement transfers out of ConsensusV2
We need to write ConsensusStreamEnded marker any time a consensus object
is transferred out of consensus. (Right now this is only done when it's
deleted or wrapped but not otherwise.)

Stacked on https://github.com/GeunhwaJeong/haneul/pull/21665

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: