Fix marker table handling in transfer-to-object for `ConsensusV2` objects
`OwnedDeleted` marker is renamed to `FastpathStreamEnded` because it is
written in many cases besides the deletion of an owned object.

We now also write the marker when a fastpath object is transferred to
consensus, so that transfer-to-object can correctly check at execution
time whether a receiving object is available.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: