[consensus_handler] Process non_randomness and randomness transactions separately
## Description 

This PR delays the merging of non randomness certificates and randomness
certificates when processing consensus transactions.
We keep them separated until we are done with assigning shared object
versions.
This should be a non-functional-change.

This will be necessary for supporting accumulators, because we need to
settle accumulators per pending checkpoint. In this case, if we are
generating two pending checkpoints (one for non randomness and one for
randomness), we need to assign versions for the accumulator object
twice.

To make things simpler, this PR also removes the randomness-specific
logic in the version manager, instead it generates a placeholder
randomness state update transaction, and let version manager to assign
versions automatically.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: