[move][tracing] Compress traces + slight format change
## Description 

Adds compression to Move traces to make them smaller. It compresses
traces using the `zstd` compression, as in benchmarking/testing this
compression provided the best compression and speed for the generated
traces.

This also updates the trace file format to be streamed. Before the trace
format was a single json object something like
```
{
  version: 1,
  events: [
    event1, 
    event2,
    ...,
    event_n,
  ]
}
```

The new trace format is now a sequence of json objects
```
{ "version": 2 }
{ event1 }
{ event2 }
...
{ event3 }
```

This also updates the representation of Move values in the generated
traces to be typed.

The base commit is the core Rust changes

The second commit are the changes to the debugger to support the new
traces format (I'm a Typescript n00b so feel free to throw things at me
here awelc).

The third commit is just updating snapshot files.

These commits can't really be broken up as they should really be landed
all together since they are both dependent on the trace format and
output.

And before you ask: yes, yes, delta encoding will be coming soon™️ 

## Test plan 

Existing tests + added a couple smaller ones. 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [X] CLI: Change in generated Move trace representation to be
compressed. Existing Move test traces will need to be regenerated in
order to me used.
- [ ] Rust SDK: