[normalized module] Restructure equality for explicit treatment of maps
## Description 

- Separates changes from  that are a bit more involved than I
would have hoped
- Adds more uses of IndexMap and explicit indications of how equality is
treated, whether it is this sort of "set based", key-value equality OR
if it should be treated as an ordered, indexed collection.

## Test plan 

- Ran tests

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: