graphql-alt: bigtable watermark
## Description

`graphql-alt`'s watermark task now fetches the latest checkpoint from
the KV store (if the service has been configured with the KV store), and
then fetches the summary for that latest checkpoint to translate the
checkpoint watermark into an epoch and transaction watermark.

## Test plan

Existing tests to make sure behaviour is correct when bigtable is not
configured. Manual testing for when bigtable is configured.