[trace-view] Updates related to trace compression
## Description 

This PR adjusts implementation of the VSCode extension to work with
compressed traces, particularly for the case of working with traces and
related meta-data produced by the replay tool.

Previously, debugging was being initiated by opening a trace file. This
had a potential to get a lot less user-friendly when dealing with
compressed traces as showing binary blob is not very meaningful and
could be quite confusing.

This PR addresses this by following the principle of fewest surprises.
When attempting to open a file the user is offered two options:
- default editor - this will still open uncompressed trace for viewing
but this option cannot be made default as the compressed trace file is
not UTF-8 encoded (in other words, VSCode will always ask to confirm
editor choice before opening the file)
- custom editor that will open uncompressed trace file and can be made a
default for opening this file type
- it a trace file is not opened but but the folder containing a trace
file (or mulitple trace files) is, the user will be offered a choice of
trace file to debug if there is more than one
- 
## Test plan 

Tested manually