graphql-alt: Object rootVersion constraint
## Description

Support fetching an object with a "root version" constraint (an
upperbound on its version). This is required to fetch dynamic fields
with in an MVCC-aware way.

In OG GraphQL, this support was added to the `owner` API, but changes to
the GraphQL schema allow us to support the query directly on the
`object` API, as well as on `multiGetObjects`.

## Test plan

New tests for the data loader:

```
haneul$ cargo nextest run -p haneul-indexer-alt-reader
```

New E2E tests for root version queries:

```
haneul$ cargo nextest run -p haneul-indexer-alt-e2e-tests
```

The query used by the data loader has been performance tested on the
production database: It was able to fetch 1000 objects in 400ms from
cold (no cache hits), and in 4ms on a warm DB.