[trace-view] Removed debug start in bytecode file plus cleanup
## Description 

This PR removes the ability to start debugging when bytecode file is
open. The main reason for this is that the existing implementation needs
to inspect all available trace files to find out which functions are
available for debugging. This is:
- quite costly as it needs to decompress ALL trace files
- not super consistent with starting to debug when a source file is open
(which selects test functions within the source file)

Additionally, this was only possible when debugging unit tests and would
not work for traced produced by replay tool, which could be potentially
quite misleading to the user. We can always bring it back in the future
if need be.

Additional cleanups are marked as comments in the PR.

## Test plan 

Tested manually that everything works