[move] Add support for summary generation to Move CLI
## Description 

Adds support for generating summaries for packages using the serialized
package summaries from the Move model 2.

A package summary can be generated one of two ways:
* Pointing it at a Move package in which case you will get a summary
with all annotations.
* Pointing it at a directory containing Move bytecode -- this will
generate summaries without any annotations.

There's future improvements to be done here in terms of removing
annotations and/or adding annotations to a model generated from
bytecode, but that's "future work" and this is a good enough MVP to get
in and start using IMO.

### One thing worth noting:

I took the view to structure the output based on package (based on
address) and then a summary for each module in its own file. This seemed
like the most reasonable/natural thing to do rather than having a file
per package, (or even only one file for all packages). But open to
feedback on whether y'all think that is the right choice or not.

## Test plan 

Added new Move CLI test.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: