Add support for dynamic transaction signing predicates
## Description 

Add support for dynamic transaction signing predicates. This allows
flexibility in terms of the types of predicates we can write and express
for the transaction signing config.

## Test plan 

Added additional tests + ran local tests to make sure we fail
immediately on an improperly formatted program + made sure it worked in
e2e scenarios.