[trace-view] Fixed breakpoints problem in no-source bytecode files
## Description 

This PR fixes a problem with setting breakpoints in disassembly files
that do not have a corresponding source file.

The problem was related to the fact that a no-source disassembly file is
treated as a source file by the runtime if no actual source file exists.
This was reflected everywhere but not in the code responsible for
checking validity of breakpoints...

## Test plan 

A test verifying that no-source disassembly file is debugged correctly
was modified. This test was previously incorrect actually (note that the
stack never showed disassembly file frame) so in addition to adding
breakpoint test, stepping through source/disassembly combination has
also been fixed.