[2/n][object runtime type tags][adapter-only] Normalize to defining ID when converting from `TypeInput` to `TypeTag`
## Description 

When converting `TypeInput`s to `TypeTag`s in the adapter also remap all
IDs to be defining-id based. This should not cause any discernible
change in behavior outside of error cases.

This adds the invariant that all typetags will be defining ID based from
the adapter down into (and including) the object runtime, with the
exception of where we explicitly want runtime IDs for typename in
natives.

## Test plan 

CI + new tests

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [X] Protocol: Normalize all type inputs to be defining ID based when
converting from `TypeInput`s in the adapter.
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------