chore: expose update_lock_file fn that does not depend on wallet context
## Description 

Walrus intends to avoid using the `HaneulClient` within the `WalletContext`
since it lacks retry semantics. To this end, we would prefer to avoid
relying on Haneul functions that force such usage. This PR extracts the
internals of `update_lock_file` so that it can be used directly without
need to pass in the `WalrusContext`.

## Test plan 

CI Pipeline.