graphql: use haneul-display to implement Display
## Description

Switch the implementation of `Display` in legacy GraphQL from a copy of
the fullnode implementation to the `haneul-display` implementation.

This is a trial run for making the same change to fullnodes.

## Test plan

E2E tests:

```
haneul$ cargo nextest run -p haneul-graphql-e2e-tests
```

Changes affect output order, (entries are now ordered by key name), and
wording for errors.