indexer-alt: retry all checkpoint fetch errors
## Description

Previously, the indexing framework would only retry a subset of remote
checkpoint fetching errors, and would wind down the indexer in response
to others.

This change makes it so that all errors (except "not found" errors) are
treated as transient and are retried with an exponential back-off. It
also simplifies the logic around cancellation, pulling it out from the
ingestion client, and introducing an outer, wrapping `select!` statement
that checks for cancellation instead.

## Test plan

Updtaed tests:

```
haneul$ cargo nextest run         \
  -p haneul-indexer-alt-framework \
  -p haneul-indexer-alt           \
  -p haneul-indexer-alt-e2e-tests
```

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: